test-requirements.txt
https://github.com/wazo-platform/wazo-test-helpers/archive/master.zip
pyhamcrest
pytest
